backend/requirements.txt
flask
flask-cors
google-genai
PyPDF2
python-dotenv
gunicorn
requests
